Return to the old return value of parseUncompressed
This reverts two changes: first, returning the size from
 `parseUncompressed` together with the data, rather than just the data;
second, returning both the name and the value of the root node as values
(`{root: <name>, value: <value>}`) rather than a key/value pair
(`{<name>: <value>}` or just `<value>` if no name).

Concerning the read offset as a return value, that definitely doesn't
belong there. If that value is required it should be exposed elsewhere
on the object rather than being tacked onto the value.