Fix parse() for compressed data in ArrayBuffers
The gzip header check assumed the data passed in could be indexed, which
is true for Node Buffer objects and Uint8Arrays but not for ArrayBuffers.

Additionally zlib.gunzip can also not deal with ArrayBuffers, so now the
data is converted to a Buffer if the type is available (it might not be in
the browser) or to a Uint8Array otherwise.